cmake: Remove Auto-Dependency system
This system did help build StreamFX for a while, but it came with its own unusual baggage. Not only was it extremely difficult to support with CI, it also becomes less and less functional with each CMake update. So in order to support future build styles, it is best to remove this system entirely. Additionally we drop support for all older versions of OBS Studio in favor of supporting OBS Studio 28.0.